Move table data source classes to separate file
WIDGETS FRNT-2 [IN PROGRESS]
Table FRNT-3 [IN PROGRESS]
