+++ BaCa2/widgets/listing/base.py
@@ -2,14 +2,12 @@ from __future__ import annotations
 
 import json
 from typing import (List, Dict, Any)
-from abc import ABC, abstractmethod
 
 from widgets.base import Widget
 from widgets.listing.columns import (Column, SelectColumn, DeleteColumn)
+from widgets.listing.data_sources import TableDataSource
 from widgets.forms import (BaCa2ModelForm, FormWidget)
 from widgets.forms.course import DeleteCourseForm
-from util.models import model_cls
-from util.models_registry import ModelsRegistry
 from main.models import Course
 
 
@@ -69,38 +67,6 @@ class TableWidget(Widget):
             'default_order_col': self.default_order_col,
             'default_order': self.default_order,
         }
-
-
-class TableDataSource(ABC):
-    @abstractmethod
-    def get_url(self) -> str:
-        raise NotImplementedError('This method has to be implemented by inheriting classes.')
-
-    @abstractmethod
-    def generate_table_widget_name(self) -> str:
-        raise NotImplementedError('This method has to be implemented by inheriting classes.')
-
-
-class ModelDataSource(TableDataSource):
-    def __init__(self, model: model_cls) -> None:
-        self.model = model
-
-    def get_url(self) -> str:
-        return f'/{self.model._meta.app_label}/models/{self.model._meta.model_name}'
-
-    def generate_table_widget_name(self) -> str:
-        return f'{self.model._meta.model_name}_table_widget'
-
-
-class CourseModelDataSource(ModelDataSource):
-    def __init__(self, model: model_cls, course: str | int | Course) -> None:
-        if not isinstance(course, str):
-            course = ModelsRegistry.get_course(course).short_name
-        self.course = course
-        super().__init__(model)
-
-    def get_url(self) -> str:
-        return f'course/{self.course}/models/{self.model._meta.model_name}'
 
 
 class TableWidgetPaging:
